[core, common] refactor entity_manager and move it to core
